feat: added traces to managed generations, not fully working yet
