+++ client/src/ee/types/types.ts
@@ -86,6 +86,47 @@ export interface AnnotationUpdate {
 }
 
 /**
+ * CommonCallParams
+ * Common parameters shared across LLM providers.
+ *
+ * Note: Each provider may handle these parameters differently or not support them at all.
+ * Please check provider-specific documentation for parameter support and behavior.
+ *
+ * Attributes:
+ *     temperature: Controls randomness in the output (0.0 to 1.0).
+ *     max_tokens: Maximum number of tokens to generate.
+ *     top_p: Nucleus sampling parameter (0.0 to 1.0).
+ *     frequency_penalty: Penalizes frequent tokens (-2.0 to 2.0).
+ *     presence_penalty: Penalizes tokens based on presence (-2.0 to 2.0).
+ *     seed: Random seed for reproducibility.
+ *     stop: Stop sequence(s) to end generation.
+ */
+export interface CommonCallParams {
+  /** Temperature */
+  temperature?: number | null;
+  /** Max Tokens */
+  max_tokens?: number | null;
+  /** Top P */
+  top_p?: number | null;
+  /** Frequency Penalty */
+  frequency_penalty?: number | null;
+  /** Presence Penalty */
+  presence_penalty?: number | null;
+  /** Seed */
+  seed?: number | null;
+  /** Stop */
+  stop?: string | string[] | null;
+}
+
+/** DependencyInfo */
+export interface DependencyInfo {
+  /** Version */
+  version: string;
+  /** Extras */
+  extras: string[] | null;
+}
+
+/**
  * EvaluationType
  * Evaluation type enum
  */
@@ -113,6 +154,135 @@ export interface Event {
   timestamp: string;
 }
 
+/**
+ * GenerationCreate
+ * Generation create model.
+ */
+export interface GenerationCreate {
+  /** Project Uuid */
+  project_uuid?: string | null;
+  /** Version Num */
+  version_num?: number | null;
+  /**
+   * Name
+   * @minLength 1
+   */
+  name: string;
+  /** Signature */
+  signature: string;
+  /** Code */
+  code: string;
+  /** Hash */
+  hash: string;
+  /** Dependencies */
+  dependencies?: Record<string, DependencyInfo>;
+  /** Arg Types */
+  arg_types?: Record<string, string>;
+  /** Archived */
+  archived?: string | null;
+  /** Custom Id */
+  custom_id?: string | null;
+  /** Prompt Template */
+  prompt_template?: string | null;
+  /** Provider */
+  provider?: string | null;
+  /** Model */
+  model?: string | null;
+  /**
+   * Common parameters shared across LLM providers.
+   *
+   * Note: Each provider may handle these parameters differently or not support them at all.
+   * Please check provider-specific documentation for parameter support and behavior.
+   *
+   * Attributes:
+   *     temperature: Controls randomness in the output (0.0 to 1.0).
+   *     max_tokens: Maximum number of tokens to generate.
+   *     top_p: Nucleus sampling parameter (0.0 to 1.0).
+   *     frequency_penalty: Penalizes frequent tokens (-2.0 to 2.0).
+   *     presence_penalty: Penalizes tokens based on presence (-2.0 to 2.0).
+   *     seed: Random seed for reproducibility.
+   *     stop: Stop sequence(s) to end generation.
+   */
+  call_params?: CommonCallParams;
+  /**
+   * Is Default
+   * @default false
+   */
+  is_default?: boolean | null;
+  /**
+   * Is Managed
+   * @default false
+   */
+  is_managed?: boolean | null;
+}
+
+/**
+ * GenerationPublic
+ * Generation public model.
+ */
+export interface GenerationPublic {
+  /** Project Uuid */
+  project_uuid?: string | null;
+  /** Version Num */
+  version_num?: number | null;
+  /**
+   * Name
+   * @minLength 1
+   */
+  name: string;
+  /** Signature */
+  signature: string;
+  /** Code */
+  code: string;
+  /** Hash */
+  hash: string;
+  /** Dependencies */
+  dependencies?: Record<string, DependencyInfo>;
+  /** Arg Types */
+  arg_types?: Record<string, string>;
+  /** Archived */
+  archived?: string | null;
+  /** Custom Id */
+  custom_id?: string | null;
+  /** Prompt Template */
+  prompt_template?: string | null;
+  /** Provider */
+  provider?: string | null;
+  /** Model */
+  model?: string | null;
+  /**
+   * Common parameters shared across LLM providers.
+   *
+   * Note: Each provider may handle these parameters differently or not support them at all.
+   * Please check provider-specific documentation for parameter support and behavior.
+   *
+   * Attributes:
+   *     temperature: Controls randomness in the output (0.0 to 1.0).
+   *     max_tokens: Maximum number of tokens to generate.
+   *     top_p: Nucleus sampling parameter (0.0 to 1.0).
+   *     frequency_penalty: Penalizes frequent tokens (-2.0 to 2.0).
+   *     presence_penalty: Penalizes tokens based on presence (-2.0 to 2.0).
+   *     seed: Random seed for reproducibility.
+   *     stop: Stop sequence(s) to end generation.
+   */
+  call_params?: CommonCallParams;
+  /**
+   * Is Default
+   * @default false
+   */
+  is_default?: boolean | null;
+  /**
+   * Is Managed
+   * @default false
+   */
+  is_managed?: boolean | null;
+  /**
+   * Uuid
+   * @format uuid
+   */
+  uuid: string;
+}
+
 /** HTTPValidationError */
 export interface HTTPValidationError {
   /** Detail */
@@ -137,6 +307,31 @@ export interface MessageParam {
   role: string;
   /** Content */
   content: (AudioPart | TextPart | ImagePart | ToolCall)[];
+}
+
+/**
+ * PlaygroundParameters
+ * Playground parameters model.
+ */
+export interface PlaygroundParameters {
+  /** Arg Values */
+  arg_values: object;
+  /** Provider name enum */
+  provider: Provider;
+  /** Model */
+  model: string;
+  generation?: GenerationCreate | null;
+}
+
+/**
+ * Provider
+ * Provider name enum
+ */
+export enum Provider {
+  OPENAI = "openai",
+  ANTHROPIC = "anthropic",
+  OPENROUTER = "openrouter",
+  GEMINI = "gemini",
 }
 
 /**
