+++ client/src/utils/generations.ts
@@ -1,25 +1,19 @@
 import api from "@/api";
-import {
-  GenerationCreate,
-  GenerationPublic,
-  GenerationUpdate,
-  PlaygroundParameters,
-} from "@/types/types";
+import { GenerationPublic, GenerationUpdate } from "@/types/types";
 import {
   queryOptions,
   useMutation,
   useQueryClient,
 } from "@tanstack/react-query";
-import { AxiosResponse } from "axios";
-import { usePostHog } from "posthog-js/react";
 
-const generationKeys = {
+export const generationKeys = {
   all: ["generations"] as const,
   lists: () => [...generationKeys.all, "list"] as const,
   list: (filters: string) => [...generationKeys.lists(), { filters }] as const,
   details: () => [...generationKeys.all, "detail"] as const,
   detail: (uuid: string) => [...generationKeys.details(), uuid] as const,
 };
+
 export const fetchGenerationsByName = async (
   generationName: string,
   projectUuid?: string
@@ -80,31 +74,6 @@ export const archiveGenerationByName = async (
   return (
     await api.delete<boolean>(
       `/projects/${projectUuid}/generations/names/${generationName}`
-    )
-  ).data;
-};
-
-export const createManagedGeneration = async (
-  projectUuid: string,
-  generationCreate: GenerationCreate
-): Promise<GenerationPublic> => {
-  return (
-    await api.post<GenerationCreate, AxiosResponse<GenerationPublic>>(
-      `/projects/${projectUuid}/managed-generations`,
-      generationCreate
-    )
-  ).data;
-};
-
-export const runGeneration = async (
-  projectUuid: string,
-  generationUuid: string,
-  playgroundValues: PlaygroundParameters
-): Promise<string> => {
-  return (
-    await api.post<Record<string, string>, AxiosResponse<string>>(
-      `/projects/${projectUuid}/generations/${generationUuid}/run`,
-      playgroundValues
     )
   ).data;
 };
@@ -188,39 +157,5 @@ export const useArchiveGenerationByNameMutation = () => {
         queryKey: generationKeys.all,
       });
     },
-  });
-};
-
-export const useCreateManagedGeneration = () => {
-  const queryClient = useQueryClient();
-  const posthog = usePostHog();
-  return useMutation({
-    mutationFn: async ({
-      projectUuid,
-      generationCreate,
-    }: {
-      projectUuid: string;
-      generationCreate: GenerationCreate;
-    }) => await createManagedGeneration(projectUuid, generationCreate),
-    onSuccess: (newVersion) => {
-      posthog.capture("managedGenerationCreated");
-      queryClient.invalidateQueries({
-        queryKey: generationKeys.list(newVersion.name),
-      });
-    },
-  });
-};
-
-export const useRunMutation = () => {
-  return useMutation({
-    mutationFn: async ({
-      projectUuid,
-      generationUuid,
-      playgroundValues,
-    }: {
-      projectUuid: string;
-      generationUuid: string;
-      playgroundValues: PlaygroundParameters;
-    }) => await runGeneration(projectUuid, generationUuid, playgroundValues),
   });
 };
