+++ client/src/utils/playground-utils.tsx
@@ -17,13 +17,8 @@ import {
   SelectTrigger,
   SelectValue,
 } from "@/components/ui/select";
-import {
-  CommonCallParams,
-  GenerationPublic,
-  PlaygroundParameters,
-  Provider,
-  UserPublic,
-} from "@/types/types";
+import { PlaygroundParameters, Provider } from "@/ee/types/types";
+import { CommonCallParams, GenerationPublic, UserPublic } from "@/types/types";
 import { userQueryOptions } from "@/utils/users";
 import { useSuspenseQuery } from "@tanstack/react-query";
 import { useEffect } from "react";
